[BE] chore: 좋아요 수 가져오기 수정

+++ backend/src/main/java/com/moyeo/entity/MoyeoTimeLine.java
@@ -52,7 +52,7 @@ public class MoyeoTimeLine extends BaseTime {
     private Boolean isTimelinePublic = true;
 
     @Column(columnDefinition = "MEDIUMINT")
-    private Integer membersCount;
+    private Long membersCount;
 
 }
